Change: ability to control opening parent ui on close
